# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 1OF1 Trader Pro Affiliate Back Office API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Public
...
  /admin/payouts:
    get:
      operationId: getAdminPayouts
      tags: [admin]
      summary: Get all payouts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Payouts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPayoutsListResponse"
    post:
      operationId: createPayout
      tags: [admin]
      summary: Create a payout batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayoutRequest"
      responses:
        "201":
          description: Payout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payout"
  /admin/payouts/{id}/mark-paid:
    post:
      operationId: markPayoutPaid
      tags: [admin]
      summary: Mark payout as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                payoutReference:
                  type: string
      responses:
        "200":
          description: Marked paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/payouts/export:
    get:
      operationId: exportPayouts
      tags: [admin]
      summary: Export payouts as CSV
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: affiliateId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string

  # Admin - Commissions
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, affiliate]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        affiliateId:
          type: integer
          nullable: true
        affiliateCode:
          type: string
          nullable: true
      required:
        - id
        - email
        - fullName
        - role
        - status
        - createdAt
...
    Payout:
      type: object
      properties:
        id:
          type: integer
        affiliateId:
          type: integer
        affiliateName:
          type: string
          nullable: true
        affiliateCode:
          type: string
          nullable: true
        totalAmount:
          type: number
        currency:
          type: string
        payoutMethod:
          type: string
          nullable: true
        payoutReference:
          type: string
          nullable: true
        periodStart:
          type: string
          format: date-time
          nullable: true
        periodEnd:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, paid, cancelled]
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - affiliateId
        - totalAmount
        - currency
        - status
        - createdAt
...
    AdminPayoutRow:
      type: object
      properties:
        id:
          type: integer
        affiliateId:
          type: integer
        totalAmount:
          type: number
        currency:
          type: string
        payoutMethod:
          type: string
          nullable: true
        payoutReference:
          type: string
          nullable: true
        periodStart:
          type: string
          format: date-time
          nullable: true
        periodEnd:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
        affiliateName:
          type: string
          nullable: true
        affiliateCode:
          type: string
          nullable: true
      required:
        - id
        - affiliateId
        - totalAmount
        - currency
        - payoutMethod
        - payoutReference
        - status
        - createdAt
        - paidAt
        - affiliateName
        - affiliateCode
    AdminPayoutsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminPayoutRow"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPending:
          type: number
        totalPaid:
          type: number
      required:
        - data
        - total
        - page
        - limit
        - totalPending
        - totalPaid
    CreatePayoutRequest:
      type: object
      properties:
        affiliateId:
          type: integer
        commissionIds:
          type: array
          items:
            type: integer
        payoutMethod:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - affiliateId
        - commissionIds
...